Various UX improments
This adds a dark theme and changes the layout to feel a bit more like GitHub,
for example, the table no longer uses a head and instead conveys the status in
text form.

This also makes the table a bit more friendly on mobile by hiding the avatars
from the owners.